+++ src/main/java/ua/kpi/iasa/taxreportingsystem/dto/LegalEntityReportDTO.java
@@ -6,6 +6,7 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import ua.kpi.iasa.taxreportingsystem.domain.User;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -16,7 +17,7 @@ import java.util.List;
 public class LegalEntityReportDTO extends ReportDTO {
     //LegalEntityReport
     private String companyName;
-    private int financialTurnover;
+    private BigDecimal financialTurnover;
     private int employeesNumber;
 
     private List<User> replacedInspectors = new ArrayList<>();
